Fix a mistake.

+++ 课外练习/14-2.c
@@ -8,11 +8,11 @@ int main() {
 	fprintf(fp,"---------------------------\n");
 	printf("---------------------------\n");
 	for(int i=1; i<=100; i++) {
-		long double t,u;
+		double t,u;
 		t=i*powl(i,2.0);
 		u=sqrtl(i);
-		printf("%6.0d%8.0Lf%9.3Lf\n",i,t,u);
-		fprintf(fp,"%6.0d%8.0Lf%9.3Lf\n",i,t,u);
+		printf("%6.0d%8.0lf%9.3lf\n",i,t,u);
+		fprintf(fp,"%6.0d%8.0lf%9.3lf\n",i,t,u);
 	}
 	fclose(fp);
 	return 0;
